+++ packages/interfaces/src/receiver.ts
@@ -5,18 +5,16 @@ import type {
   PubSubTopic,
   Unsubscribe
 } from "./misc.js";
-import type { Callback, ProtocolOptions } from "./protocols.js";
+import type { Callback } from "./protocols.js";
 
 export type ActiveSubscriptions = Map<PubSubTopic, ContentTopic[]>;
 
 export interface IReceiver {
   toSubscriptionIterator: <T extends IDecodedMessage>(
-    decoders: IDecoder<T> | IDecoder<T>[],
-    opts?: ProtocolOptions
+    decoders: IDecoder<T> | IDecoder<T>[]
   ) => Promise<IAsyncIterator<T>>;
   subscribe: <T extends IDecodedMessage>(
     decoders: IDecoder<T> | IDecoder<T>[],
-    callback: Callback<T>,
-    opts?: ProtocolOptions
+    callback: Callback<T>
   ) => Unsubscribe | Promise<Unsubscribe>;
 }