+++ packages/interfaces/src/sender.ts
@@ -1,10 +1,6 @@
 import type { IEncoder, IMessage } from "./message.js";
-import type { ProtocolOptions, SendResult } from "./protocols.js";
+import type { SendResult } from "./protocols.js";
 
 export interface ISender {
-  send: (
-    encoder: IEncoder,
-    message: IMessage,
-    opts?: ProtocolOptions
-  ) => Promise<SendResult>;
+  send: (encoder: IEncoder, message: IMessage) => Promise<SendResult>;
 }