+++ packages/utils/src/common/index.ts
@@ -1,5 +1,8 @@
 export * from "./is_defined.js";
 export * from "./random_subset.js";
+export * from "./group_by.js";
+export * from "./to_async_iterator.js";
+export * from "./is_size_valid.js";
 
 export function removeItemFromArray(arr: unknown[], value: unknown): unknown[] {
   const index = arr.indexOf(value);
@@ -8,6 +11,3 @@ export function removeItemFromArray(arr: unknown[], value: unknown): unknown[] {
   }
   return arr;
 }
-export * from "./group_by.js";
-export * from "./to_async_iterator.js";
-export * from "./is_size_valid.js";